chore: create lib/api
