+++ src/features/auth/infrastructure/getUser.ts
@@ -1,13 +1,1 @@
-import { omit } from "ramda";
-
-import type { IUser } from "@/features/auth/types/IUser";
-import { httpService } from "@/lib/http";
-
-import type { IUserDto } from "./types/IUserDto";
-
-export const getUser = () => {
-  // mocking current user and its cartId by passing id=1
-  return httpService
-    .get<IUserDto>("users/1")
-    .then((res) => ({ ...(omit(["password"], res) as IUser), cartId: 1 }));
-};
+export { getUser } from "@/lib/api/auth/users/{user-id}/user-query";
