--- src/features/auth/infrastructure/types/IUserDto.ts
@@ -1,5 +1,0 @@
-import type { IUser } from "@/features/auth/types/IUser";
-
-export interface IUserDto extends IUser {
-  password: string;
-}
