--- src/features/carts/infrastructure/types/ICartDto.ts
@@ -1,3 +1,0 @@
-import type { ICart } from "@/features/carts/types/ICart";
-
-export interface ICartDto extends ICart {}
