--- src/features/carts/infrastructure/types/ICartProductDto.ts
@@ -1,3 +1,0 @@
-import type { ICartProduct } from "@/features/carts/types/ICartProduct";
-
-export interface ICartProductDto extends ICartProduct {}
