+++ src/features/carts/infrastructure/usePurchase.ts
@@ -1,18 +1,1 @@
-import { useState } from "react";
-
-// todo: use msw to mock this feature
-export const usePurchase = () => {
-  const [isLoading, setIsLoading] = useState(false);
-
-  const purchase = async () => {
-    return new Promise<void>((resolve) => {
-      setIsLoading(true);
-      setTimeout(() => {
-        resolve();
-        setIsLoading(false);
-      }, 400);
-    });
-  };
-
-  return [purchase, isLoading] as const;
-};
+export { usePurchaseMutation as usePurchase } from "@/lib/api/carts/{cart-id}/purchase-command";
