+++ src/features/carts/types/ICartProduct.ts
@@ -1,6 +1,1 @@
-// eslint-disable-next-line import/no-restricted-paths
-import type { IProduct } from "@/features/products/types/IProduct";
-
-export interface ICartProduct extends IProduct {
-  quantity: number;
-}
+export type { CartProductDto as ICartProduct } from "@/lib/api/carts/{cart-id}/cart-product-dto";
