--- src/features/products/infrastructure/types/IProductDto.ts
@@ -1,3 +1,0 @@
-import type { IProduct } from "@/features/products/types/IProduct";
-
-export interface IProductDto extends IProduct {}
