+++ src/features/products/types/Category.ts
@@ -1,6 +1,1 @@
-export enum Category {
-  Men_clothing = "men's clothing",
-  Women_clothing = "women's clothing",
-  Jewelery = "jewelery",
-  Electronics = "electronics",
-}
+export { Category } from "@/lib/api/products/{product-id}/product-dto";
