+++ src/lib/api/auth/login/login-command.ts
@@ -1,0 +1,7 @@
+import { httpService } from "@/lib/http";
+
+import type { LoginDto } from "./login-dto";
+
+export const loginUser = (body: LoginDto) => {
+  return httpService.post<string>("auth/login", body);
+};
