+++ src/lib/api/auth/login/login-dto.ts
@@ -1,0 +1,4 @@
+export interface LoginDto {
+  username: string;
+  password: string;
+}
