+++ src/lib/api/auth/users/{user-id}/user-dto.ts
@@ -1,0 +1,24 @@
+export interface IAddress {
+  city: string;
+  street: string;
+  number: number;
+  zipcode: string;
+  geolocation: {
+    lat: string;
+    long: string;
+  };
+}
+
+export interface UserDto {
+  id: number;
+  email: string;
+  username: string;
+  name: {
+    firstname: string;
+    lastname: string;
+  };
+  phone: string;
+  address: IAddress;
+  cartId: number;
+  password: string;
+}
