+++ src/lib/api/carts/{cart-id}/cart-dto.ts
@@ -1,0 +1,11 @@
+interface IProduct {
+  productId: number;
+  quantity: number;
+}
+
+export interface CartDto {
+  id: number;
+  userId: number;
+  date: string;
+  products: IProduct[];
+}
