+++ src/lib/api/carts/{cart-id}/cart-product-dto.ts
@@ -1,0 +1,5 @@
+import type { ProductDto } from "../../products/{product-id}/product-dto";
+
+export interface CartProductDto extends ProductDto {
+  quantity: number;
+}
