+++ src/lib/api/carts/{cart-id}/clear-cart-command.ts
@@ -1,0 +1,33 @@
+import { useMutation } from "@tanstack/react-query";
+
+import { httpService } from "@/lib/http";
+import { Logger } from "@/lib/logger";
+
+interface IClearCartValues {
+  cartId: number;
+}
+
+export const useClearCartMutation = () => {
+  const { mutateAsync, isLoading } = useMutation<
+    void,
+    unknown,
+    IClearCartValues
+  >((body) => httpService.delete(`carts/${body.cartId}`));
+
+  const handler = (body: IClearCartValues) => {
+    return mutateAsync(body)
+      .then(async () => {
+        // optionally mutate related data
+      })
+      .catch((e) => {
+        // listen for a specific error and act respectively (e.g. throwing a specific error and catch it later)
+
+        // notify backend about the error if needed
+        Logger.error("An error occurred during clearing the cart", e as Error);
+
+        throw e;
+      });
+  };
+
+  return [handler, isLoading] as const;
+};
