+++ src/lib/api/products/products-list/products-list-dto.ts
@@ -1,0 +1,8 @@
+import type { IMeta } from "@/types/IMeta";
+
+import type { ProductDto } from "../{product-id}/product-dto";
+
+export interface ProductsListDto {
+  products: ProductDto[];
+  meta: IMeta;
+}
