+++ src/lib/components/Layout/Navbar/index.tsx
@@ -12,6 +12,7 @@ import {
   useDisclosure,
 } from "@chakra-ui/react";
 
+// eslint-disable-next-line import/no-restricted-paths
 import { useAuthStore } from "@/features/auth/application/authStore";
 import { useNotImplementedYetToast } from "@/lib/components/Toast/useNotImplementedYetToast";
 import { Link, useNavigate } from "@/lib/router";
