+++ src/lib/components/Layout/Navbar/useNavItems.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/no-restricted-paths
 import { useAuthStore } from "@/features/auth/application/authStore";
 
 import type { INavItem } from "./INavItem";
